// src/frontend/src/backend.d.ts
export interface ChatMessage {
    id: string;
    senderIsOwner: boolean;
    messageText: string;
    isReply: boolean;
    originalMessageId?: string;
    replyToMessageId?: string;
    timestamp: bigint;
    senderName: string;
    senderEmail: string;
}